Test refactoring to specific tests per integration point
