+++ src/test/java/io/github/pmckeown/rest/client/GetMetricsTest.java
@@ -1,0 +1,16 @@
+package io.github.pmckeown.rest.client;
+
+import org.junit.Test;
+
+/**
+ * Test metrics retrieval
+ *
+ * @author Paul McKeown
+ */
+public class GetMetricsTest extends AbstractDependencyTrackIntegrationTest {
+
+    @Test
+    public void thatMetricsCanBeRetrieved() throws Exception {
+        
+    }
+}
